dddddddddd

+++ src/main/java/com/apec/pos/Dto/TypeDto/TypeResponseAdmin.java
@@ -6,13 +6,13 @@ import com.apec.pos.entity.TypeFoodEntity;
 
 public class TypeResponseAdmin {
 
-	private Integer totalRow;
+	private Long totalRow;
 	private List<TypefoodResponseData> data;
 	
-	public Integer getTotalRow() {
+	public Long getTotalRow() {
 		return totalRow;
 	}
-	public void setTotalRow(Integer totalRow) {
+	public void setTotalRow(Long totalRow) {
 		this.totalRow = totalRow;
 	}
 	public List<TypefoodResponseData> getData() {
